fix(core): expose only component registry to public API
And hide internal components DI token
